refactor: ホットキー関連の型を新しいドメインファイルに移動
